Statically initialize built-in op levels
This unifies all the built-in operator data in one place and statically initializes it all in one place. The culmination of this patch series thus removes the necessity of calling two functions that modify global static variables before you're allowed to call the parser, in line with the aims of #891.
